+++ examples/Cargo.toml
@@ -30,6 +30,7 @@ members = [
     "panicker",
     "ping",
     "ping-gas",
+    "program_generator",
     "program-id",
     "rwlock",
     "rwlock_write",
